% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-relationship-dates.R
\name{clean_lei_relationship_dates}
\alias{clean_lei_relationship_dates}
\title{Clean LEI relationships}
\usage{
clean_lei_relationship_dates(raw_lei_relationships)
}
\description{
Find the relationship period dates and convert them into a useful format
}
