% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr-package.R
\docType{package}
\name{gleifr-package}
\alias{gleifr}
\alias{gleifr-package}
\title{gleifr: Client for the 'GLEIF' API}
\description{
Download legal entity reference data from the 'Global Legal Entity Identifier Foundation' ('GLEIF') API. Retrieve Legal Entity Identifier ('LEI') records, their direct and ultimate parent and child relationships, accredited issuers ('Local Operating Units'), and mappings from 'LEI' codes to other identifiers such as 'ISIN', 'BIC', and 'MIC'. See \url{https://www.gleif.org/en/lei-data/gleif-api} for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/gleifr/}
  \item \url{https://github.com/m-muecke/gleifr}
  \item Report bugs at \url{https://github.com/m-muecke/gleifr/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

Authors:
\itemize{
  \item Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})
}

}
\keyword{internal}
