% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_children}
\alias{lei_children}
\title{Fetch LEI child records}
\usage{
lei_children(id, type = c("direct", "ultimate"), limit = 200L, simplify = TRUE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch the children for.}

\item{type}{(\code{character(1)})\cr
The type of children to fetch. One of \code{"direct"} or \code{"ultimate"}. Default is \code{"direct"}.}

\item{limit}{(\code{NULL} | \code{integer(1)})\cr
Maximum number of records to return. Default \code{200L}. Use \code{NULL} to fetch all.}

\item{simplify}{(\code{logical(1)})\cr
Should the output be simplified? Default \code{TRUE}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{name}: The attribute name
\item \strong{value}: The attribute value
}

When \code{simplify = FALSE}, a \code{list()} of the raw record objects from the API.

When \code{simplify = TRUE} and no records match, \code{NULL}.
}
\description{
Fetches the direct or ultimate child records of a given LEI.
}
\examples{
\donttest{
# get direct children
head(lei_children("O2RNE8IBXP4R0TD8PU41", limit = 5))

# get ultimate children
head(lei_children("O2RNE8IBXP4R0TD8PU41", type = "ultimate", limit = 5))
}
}
\seealso{
\code{\link[=lei_parent]{lei_parent()}} to fetch the parent record of a LEI.
}
