% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_countries}
\alias{lei_countries}
\title{Fetch the list of countries}
\usage{
lei_countries()
}
\value{
A \code{data.frame()} with columns:
\itemize{
\item \strong{code}: The country code
\item \strong{name}: The country name
}
}
\description{
Fetches the list of countries (ISO 3166-1) recognized by the GLEIF API.
}
\examples{
\donttest{
lei_countries()
}
}
