% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_isins}
\alias{lei_isins}
\title{Fetch ISINs for a LEI}
\usage{
lei_isins(id)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch ISINs for.}
}
\value{
A \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{isin}: The ISIN
}
}
\description{
Fetches the ISINs associated with a given LEI.
}
\examples{
\dontrun{
lei_isins("529900W18LQJJN6SJ336")
}
}
