% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_jurisdictions}
\alias{lei_jurisdictions}
\title{Fetch the list of jurisdictions}
\usage{
lei_jurisdictions()
}
\value{
A \code{data.frame()} with columns:
\itemize{
\item \strong{code}: The jurisdiction code
\item \strong{name}: The jurisdiction name
}
}
\description{
Fetches the list of jurisdictions (countries and their subdivisions) recognized by the GLEIF API.
}
\examples{
\dontrun{
lei_jurisdictions()
}
}
