% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_mapping}
\alias{lei_mapping}
\title{Download the latest LEI mapping data}
\source{
\url{https://www.gleif.org/en/lei-data/lei-mapping}
}
\usage{
lei_mapping(type = c("isin", "bic", "mic", "oc"))
}
\arguments{
\item{type}{(\code{character(1)})\cr
The type of mapping data to download.
One of \code{"isin"}, \code{"bic"}, \code{"mic"}, or \code{"oc"}. Default is \code{"isin"}.}
}
\value{
A \code{data.frame()} with the lei and the corresponding mapping.
}
\description{
Download the latest Legal Entity Identifier (LEI) mapping data from the
Global Legal Entity Identifier Foundation (GLEIF).
}
\examples{
\donttest{
lei_mapping("isin")
}
}
