% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_parent}
\alias{lei_parent}
\title{Fetch a LEI parent record}
\usage{
lei_parent(id, type = c("direct", "ultimate"), simplify = TRUE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch the parent for.}

\item{type}{(\code{character(1)})\cr
The type of parent to fetch. One of \code{"direct"} or \code{"ultimate"}. Default is \code{"direct"}.}

\item{simplify}{(\code{logical(1)})\cr
Should the output be simplified? Default \code{TRUE}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{name}: The attribute name
\item \strong{value}: The attribute value
}

When \code{simplify = FALSE}, a named \code{list()} containing the raw API response.
}
\description{
Fetches the direct or ultimate parent record of a given LEI.
}
\examples{
\donttest{
# get direct parent
head(lei_parent("529900W18LQJJN6SJ336"))

# get ultimate parent
head(lei_parent("529900W18LQJJN6SJ336", type = "ultimate"))
}
}
\seealso{
\code{\link[=lei_children]{lei_children()}} to fetch the child records of a LEI.
}
