% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_parents}
\alias{lei_parents}
\title{Fetch LEI parent records}
\usage{
lei_parents(id, type = c("direct", "ultimate"), simplify = TRUE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch the parent for.}

\item{type}{(\code{character(1)})\cr
The type of parent to fetch. One of \code{"direct"} or \code{"ultimate"}. Default is \code{"direct"}.}

\item{simplify}{(\code{logical(1)})\cr
Should the output be simplified? Default \code{TRUE}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{name}: The attribute name
\item \strong{value}: The attribute value
}

When \code{simplify = FALSE}, a named \code{list()} containing the raw API response.
}
\description{
Fetches the direct or ultimate parent record of a given LEI.
}
\examples{
\dontrun{
# get direct parent
lei_parents("529900W18LQJJN6SJ336")

# get ultimate parent
lei_parents("529900W18LQJJN6SJ336", type = "ultimate")
}
}
