% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_record_by_id}
\alias{lei_record_by_id}
\title{Fetch a LEI record}
\usage{
lei_record_by_id(id, simplify = TRUE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch.}

\item{simplify}{(\code{logical(1)})\cr
Should the output be simplified? Default \code{TRUE}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{name}: The attribute name
\item \strong{value}: The attribute value
}

When \code{simplify = FALSE}, a named \code{list()} containing the raw API response.
}
\description{
Fetch a single LEI record by its Legal Entity Identifier. To fetch records by attribute filters,
see \code{\link[=lei_records]{lei_records()}}.
}
\examples{
\donttest{
# get simplified long-format data.frame
head(lei_record_by_id("529900W18LQJJN6SJ336"))

# get raw API response as named list
str(lei_record_by_id("529900W18LQJJN6SJ336", simplify = FALSE), max.level = 2)
}
}
\seealso{
\code{\link[=lei_records]{lei_records()}} to fetch records by attribute filters.
}
