% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_regions}
\alias{lei_regions}
\title{Fetch LEI regions}
\usage{
lei_regions()
}
\value{
A \code{data.frame()} with columns:
\itemize{
\item \strong{code}: The region code
\item \strong{language}: The language of the region name
\item \strong{name}: The region name
}
}
\description{
Fetches the list of regions from the GLEIF API.
}
\examples{
\donttest{
head(lei_regions())
}
}
