% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-timestamps.R
\name{parse_lei_unix_ts}
\alias{parse_lei_unix_ts}
\title{Parse LEI Unix Timestamps}
\usage{
parse_lei_unix_ts(ts)
}
\arguments{
\item{ts}{Unix millisecond timestamps}
}
\value{
POSIXct timestamps
}
\description{
LEI data has some timestamps in Unix millisecond timestamps. This functions parses them into POSIXct.
}
