% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultimate-parent.R
\name{ultimate_parent}
\alias{ultimate_parent}
\title{Ultimate Parent}
\usage{
ultimate_parent(entities, relationships)
}
\arguments{
\item{entities}{The output of \link{read_gleif_entities_csv}}

\item{relationships}{The output of \link{read_gleif_relationships_csv}}
}
\value{
The entities data frame with two columns appended. The LEI and legal name of the ultimate parent.
}
\description{
Finds the ultimate owner of each entity. This is done by building a graph of a relationships, then finding the node with zero outward degree in each connected subgraph
}
