% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\alias{lei_cache_dir}
\alias{lei_cache_clear}
\title{Get or manage the gleifr API cache}
\usage{
lei_cache_dir()

lei_cache_clear()
}
\value{
\code{lei_cache_dir()} returns a single string giving the path to the cache directory.

\code{lei_cache_clear()} is called for its side effect of deleting the cached responses and returns
\code{NULL} invisibly.
}
\description{
\code{lei_cache_dir()} returns the path where cached API responses are stored.
\code{lei_cache_clear()} clears all cached responses.
}
\details{
The cache is only used when enabled with \code{options(gleifr.cache = TRUE)}.
Cached responses are stored for 1 day by default, but this can be customized with
\code{options(gleifr.cache_max_age = seconds)}.
}
\examples{
\dontrun{
# enable caching
options(gleifr.cache = TRUE)

# view cache location
lei_cache_dir()

# clear the cache
lei_cache_clear()
}
}
