% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr-package.R
\docType{package}
\name{gleifr-package}
\alias{gleifr}
\alias{gleifr-package}
\title{gleifr: GLEIF API Wrapper}
\description{
Download data from GLEIF API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/gleifr}
  \item \url{https://github.com/m-muecke/gleifr}
  \item Report bugs at \url{https://github.com/m-muecke/gleifr/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

Authors:
\itemize{
  \item Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})
}

}
\keyword{internal}
