% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleifr.R
\name{lei_children}
\alias{lei_children}
\title{Fetch LEI child records}
\usage{
lei_children(id, simplify = TRUE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The Legal Entity Identifier (LEI) to fetch the children for.}

\item{simplify}{(\code{logical(1)})\cr
Should the output be simplified? Default \code{TRUE}.}
}
\value{
When \code{simplify = TRUE}, a long-format \code{data.frame()} with columns:
\itemize{
\item \strong{lei}: The Legal Entity Identifier
\item \strong{name}: The attribute name
\item \strong{value}: The attribute value
}

When \code{simplify = FALSE}, a named \code{list()} containing the raw API response.
}
\description{
Fetches the direct child records of a given LEI.
}
\examples{
\dontrun{
lei_children("529900W18LQJJN6SJ336")
}
}
