% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-lei-csv.R
\name{read_gleif_relationships_csv}
\alias{read_gleif_relationships_csv}
\title{Read GLEIF relationships csv file}
\usage{
read_gleif_relationships_csv(file)
}
\description{
Read GLEIF relationships csv file
}
